requirements.txt
# Core
numpy>=1.20.0
opencv-python>=4.5.0
pyyaml>=6.0

# rtmlib (DWPose)
rtmlib>=0.0.10

# onnxruntime-gpu는 Dockerfile에서 별도 설치
# onnxruntime-gpu  # 주석 처리